Reorganize the project architecture Move the controllers in its feature's namespace Use annotations for routing
