--- back/app/config/routing/home.yml
@@ -1,27 +1,0 @@
-# All actions not secured
-
-hello_world:
-    path: /
-    defaults: { _controller: 'CiTron:Default:hello' }
-
-user_registration:
-    path: /registration
-    defaults: { _controller: 'CiTron:User\Registration:createUser' }
-    methods: [POST]
-
-login_check:
-    path: /login
-    methods: [POST]
-
-login_status:
-    path: /login-status.json
-    defaults: { _controller: 'CiTron:User\Login:status' }
-    methods: [GET]
-
-logout:
-    path: /logout
-    methods: [GET]
-
-logout_success:
-    path: /logout-success.json
-    defaults: { _controller: 'CiTron:User\Login:logoutSuccess' }
